refactor: section 리스트가 아닌 station id로 station들을 조회하도록 변경
