+++ src/main/java/wooteco/subway/dao/StationDao.java
@@ -2,7 +2,6 @@ package wooteco.subway.dao;
 
 import java.util.List;
 import java.util.Optional;
-import wooteco.subway.domain.Section;
 import wooteco.subway.domain.Station;
 
 public interface StationDao {
@@ -13,7 +12,7 @@ public interface StationDao {
 
     List<Station> findAll();
 
-    List<Station> findAllBySection(List<Section> sections);
+    List<Station> findAllByIds(List<Long> stationIds);
 
     int deleteById(Long id);
 }
