+++ src/test/java/wooteco/subway/service/fake/FakeStationDao.java
@@ -6,12 +6,9 @@ import java.lang.reflect.Field;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
-import java.util.Set;
-import java.util.TreeSet;
 import org.springframework.dao.DuplicateKeyException;
 import org.springframework.util.ReflectionUtils;
 import wooteco.subway.dao.StationDao;
-import wooteco.subway.domain.Section;
 import wooteco.subway.domain.Station;
 
 public class FakeStationDao implements StationDao {
@@ -45,21 +42,10 @@ public class FakeStationDao implements StationDao {
     }
 
     @Override
-    public List<Station> findAllBySection(List<Section> sections) {
-        List<Long> stationIds = getStationIds(sections);
-
+    public List<Station> findAllByIds(List<Long> stationIds) {
         return stations.stream()
                 .filter(station -> stationIds.contains(station.getId()))
                 .collect(toList());
-    }
-
-    private List<Long> getStationIds(List<Section> sections) {
-        Set<Long> stationIds = new TreeSet<>();
-        for (Section section : sections) {
-            stationIds.add(section.getUpStationId());
-            stationIds.add(section.getDownStationId());
-        }
-        return new ArrayList<>(stationIds);
     }
 
     @Override
